Split up node logic
